moving splines.py to utils submodule + some minor changes to readme
