+++ src/dynode/simulation/__init__.py
@@ -1,12 +1,8 @@
 """Module to hold all DynODE ODE flows."""
 
 from .odes import AbstractODEParams, simulate
-from .splines import base_equation, conditional_knots, evaluate_cubic_spline
 
 __all__ = [
     "simulate",
     "AbstractODEParams",
-    "evaluate_cubic_spline",
-    "base_equation",
-    "conditional_knots",
 ]
