+++ src/dynode/utils/__init__.py
@@ -2,7 +2,7 @@
 
 import logging
 
-from . import log
+from . import log, vis_utils
 from .custom_log_formatter import CustomLogFormatter
 from .datetime_utils import (
     date_to_epi_week,
@@ -11,19 +11,20 @@ from .datetime_utils import (
     sim_day_to_epiweek,
 )
 from .log_decorator import log_decorator
+from .splines import base_equation, conditional_knots, evaluate_cubic_spline
 from .utils import (
     drop_keys_with_substring,
     flatten_list_parameters,
     identify_distribution_indexes,
     vectorize_objects,
 )
-from .vis_utils import *  # noqa: F403
 
 # Fetching the global logger called dynode
 logger = logging.getLogger("dynode")
 
 __all__ = [
     "log",
+    "vis_utils",
     "log_decorator",
     "CustomLogFormatter",
     "logger",
@@ -35,4 +36,7 @@ __all__ = [
     "flatten_list_parameters",
     "drop_keys_with_substring",
     "identify_distribution_indexes",
+    "base_equation",
+    "conditional_knots",
+    "evaluate_cubic_spline",
 ]
