+++ src/dynode/utils/splines.py
